Refactor dashboard into modules
