+++ dashboard/src/BreadcrumbsDashboard.css
@@ -1,0 +1,4 @@
+@keyframes spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
